protocol: delete dead contract surfaces and consolidate module ceremony (deslop PR-1)
Deletions (zero production consumers, verified by repo-wide grep + gates):
- src/execution/ + src/worker-bootstrap/: Execution.Request/Result and
  bootstrap payloads — the process transport speaks its own two-line wire;
  no caller constructs these contracts.
- src/command/: Command request vocabulary owned by the deleted dispatch
  runtime; wait matcher dispatchEvidence + channels harness updated.
- ipc method table: coordinator.*/worker.* rows (11 methods) — only the
  machine.attach/run_cell/call_tool wire is spoken; ipc tests re-pointed.
- ledger command stream + CommandAuthorized/CommandDenied fact schemas
  (dormant, no producer ever shipped) + producer-manifest row.
- work-item dormant #493 vocab: CacheKey/ReplayKey/NondeterminismManifest/
  NondeterminismCategory/EnvironmentSubsetField/cacheKeyOf (schema-only,
  no lookup implementation; EnvironmentFingerprint + reusedFromAttemptId
  stay — ledger lifecycle consumes them).
- dead barrel aliases: Wait.OwnerKind, Wait.effectiveThreshold,
  Engagement.Terms, Engagement.isTerminal (fold helper deleted too).

Consolidations (same contracts, less ceremony):
- policy point files 4 → 1: policy-point.ts now owns contract vocabulary,
  registry, and input validators; cross-file Module namespaces dropped.
- storage/ 7 files → 1: single Storage namespace, declare-module
  augmentation ceremony dropped (type-only surface, ledger unaffected).
- ingress target/surface-key/payload folds merged into ingress/index.ts.
- policy/decision.ts inlined into the PolicyDecision namespace.

Baselines shrink only: schema-snapshot (16 type keys removed),
lint-tools vocab ratchet (execution, worker-bootstrap unmapped rows),
naming grandfathers for deleted files.
