+++ packages/protocol/src/index.ts
@@ -13,14 +13,9 @@ export * from "./actor/index.js";
 export * from "./communication/index.js";
 export * from "./gateway/index.js";
 export * from "./ingress/index.js";
-export * from "./ingress/target.js";
-export * from "./ingress/surface-key.js";
-export * from "./ingress/payload.js";
 export * from "./policy/index.js";
 export * from "./artifact/index.js";
-export * from "./execution/index.js";
 export * from "./ipc/index.js";
-export * from "./worker-bootstrap/index.js";
 export * from "./work-item/index.js";
 export * from "./deadline/index.js";
 export * from "./wait/index.js";
@@ -32,5 +27,4 @@ export * from "./ledger/index.js";
 export * from "./cron/index.js";
 export * from "./storage/index.js";
 export * from "./trace/index.js";
-export * from "./command/index.js";
 export * from "./app-connector/index.js";
