--- packages/protocol/src/policy/point-registry.ts
@@ -1,206 +1,0 @@
-import { PolicyPointContractModule } from "./point-contract.js";
-
-export namespace PolicyPointRegistryModule {
-  const contract = PolicyPointContractModule.contract;
-  const preBoundary = PolicyPointContractModule.preBoundary;
-  const postBoundary = PolicyPointContractModule.postBoundary;
-  type RegisteredPolicyPointId = PolicyPointContractModule.RegisteredPolicyPointId;
-  type PolicyPointContract = PolicyPointContractModule.PolicyPointContractSnapshot;
-
-  export const PolicyPointRegistry = Object.freeze({
-    "dispatch.action.pre": contract(
-      "dispatch.action.pre",
-      "pre",
-      ["dispatch"],
-      ["actor", "dispatchId", "action", "target"],
-      ["audit.annotate", "run.abort"],
-      ...preBoundary,
-    ),
-    "run.lifecycle.pre": contract(
-      "run.lifecycle.pre",
-      "pre",
-      ["run"],
-      ["actorId", "sessionId", "runId"],
-      [
-        "audit.annotate",
-        "run.abort",
-        "delegation.set_constraints",
-        "prompt.append_context",
-        "prompt.inject_message",
-      ],
-      ...preBoundary,
-    ),
-    "run.turn.pre": contract(
-      "run.turn.pre",
-      "pre",
-      ["run"],
-      ["sessionId", "runId", "turnIndex"],
-      [
-        "audit.annotate",
-        "run.abort",
-        "run.retry_after",
-        "prompt.append_context",
-        "prompt.inject_message",
-      ],
-      ...preBoundary,
-    ),
-    "prompt.context.pre": contract(
-      "prompt.context.pre",
-      "pre",
-      ["prompt"],
-      ["sessionId", "runId", "turnIndex"],
-      ["prompt.append_context", "prompt.inject_message", "prompt.replace", "audit.annotate"],
-      // Prompt context policy is advisory: failure should not block the run,
-      // but it still marks a side-effect boundary because it can rewrite prompt input.
-      "fail-open",
-      true,
-    ),
-    "tool.catalog.pre": contract(
-      "tool.catalog.pre",
-      "pre",
-      ["tool"],
-      ["sessionId", "runId", "availableTools"],
-      ["tool.filter", "audit.annotate", "run.abort"],
-      ...preBoundary,
-    ),
-    "connection.llm.pre": contract(
-      "connection.llm.pre",
-      "pre",
-      ["connection"],
-      ["sessionId", "runId", "modelId"],
-      [
-        "prompt.append_context",
-        "prompt.inject_message",
-        "run.abort",
-        "audit.annotate",
-        "model.override",
-      ],
-      ...preBoundary,
-    ),
-    "connection.llm.post": contract(
-      "connection.llm.post",
-      "post",
-      ["connection"],
-      ["sessionId", "runId", "modelId", "responseTokens"],
-      ["audit.annotate", "run.abort", "prompt.inject_message", "run.replace_messages"],
-      ...postBoundary,
-    ),
-    "tool.native.pre": contract(
-      "tool.native.pre",
-      "pre",
-      ["tool"],
-      ["sessionId", "runId", "toolId", "toolInput"],
-      [
-        "tool.filter",
-        "tool.rewrite_input",
-        "tool.skip_invocation",
-        "tool.require_approval",
-        "run.abort",
-        "audit.annotate",
-      ],
-      ...preBoundary,
-    ),
-    "tool.mcp.pre": contract(
-      "tool.mcp.pre",
-      "pre",
-      ["tool"],
-      ["sessionId", "runId", "toolId", "mcpServerId", "toolInput"],
-      [
-        "tool.filter",
-        "tool.rewrite_input",
-        "tool.skip_invocation",
-        "tool.require_approval",
-        "run.abort",
-        "audit.annotate",
-      ],
-      ...preBoundary,
-    ),
-    "delegation.worker.pre": contract(
-      "delegation.worker.pre",
-      "pre",
-      ["worker"],
-      ["sessionId", "runId", "workerRunId", "workerProfile"],
-      ["delegation.set_constraints", "delegation.require_approval", "run.abort", "audit.annotate"],
-      ...preBoundary,
-    ),
-    "tool.native.post": contract(
-      "tool.native.post",
-      "post",
-      ["tool"],
-      ["sessionId", "runId", "toolId", "toolResult"],
-      ["audit.annotate", "run.abort", "tool.rewrite_output"],
-      ...postBoundary,
-    ),
-    "tool.mcp.post": contract(
-      "tool.mcp.post",
-      "post",
-      ["tool"],
-      ["sessionId", "runId", "toolId", "mcpServerId", "toolResult"],
-      ["audit.annotate", "run.abort", "tool.rewrite_output"],
-      ...postBoundary,
-    ),
-    "delegation.worker.post": contract(
-      "delegation.worker.post",
-      "post",
-      ["worker"],
-      ["sessionId", "runId", "workerRunId", "workerResult"],
-      ["audit.annotate"],
-      ...postBoundary,
-    ),
-    "run.turn.post": contract(
-      "run.turn.post",
-      "post",
-      ["run"],
-      ["sessionId", "runId", "turnIndex", "turnResult"],
-      [
-        "audit.annotate",
-        "run.abort",
-        "run.continue_with_prompt",
-        "prompt.inject_message",
-        "run.replace_messages",
-      ],
-      ...postBoundary,
-    ),
-    "run.completion.pre": contract(
-      "run.completion.pre",
-      "pre",
-      ["run"],
-      ["sessionId", "runId", "completionCandidate"],
-      ["audit.annotate", "run.abort", "prompt.append_context", "run.replace_messages"],
-      ...preBoundary,
-    ),
-    "work.complete.pre": contract(
-      "work.complete.pre",
-      "pre",
-      ["work"],
-      [
-        "workItemHash",
-        "requestId",
-        "contractRevision",
-        "basisRef",
-        "expectedHead",
-        "completionCandidate",
-        "unresolvedBlockerIds",
-      ],
-      ["audit.annotate", "run.abort", "work.allow_asserted"],
-      ...preBoundary,
-    ),
-    "run.lifecycle.post": contract(
-      "run.lifecycle.post",
-      "post",
-      ["run"],
-      ["sessionId", "runId", "runOutcome"],
-      ["audit.annotate"],
-      ...postBoundary,
-    ),
-    "run.error.error": contract(
-      "run.error.error",
-      "error",
-      ["run"],
-      ["sessionId", "runId", "errorCode", "errorPhase"],
-      ["audit.annotate", "run.abort", "run.retry_after"],
-      "fail-closed",
-      false,
-    ),
-  } satisfies Record<RegisteredPolicyPointId, PolicyPointContract>);
-}
