--- packages/protocol/src/storage/namespace.ts
@@ -1,2 +1,0 @@
-// This file owns the shared type-only namespace contract for storage domains.
-export namespace Storage {}
