+++ packages/protocol/src/tool/index.ts
@@ -150,10 +150,8 @@ export namespace Tool {
   export type Result = z.infer<typeof Result>;
 
   /**
-   * #500 C4: tool-catalog selection vocabulary, folded in from the deleted
-   * `tool-selection/` sibling — one noun namespace for the tool grammar.
-   * Consumed by the openomni tool catalog resolver and
-   * `WorkerBootstrap.AgentDefinition.tools`.
+   * Tool-catalog selection vocabulary — one noun namespace for the tool
+   * grammar, consumed by the openomni tool catalog resolver.
    */
   export const Category = z.enum(["filesystem", "execution", "delegation", "mcp", "custom"]);
   export type Category = z.infer<typeof Category>;
