+++ packages/protocol/src/work-item/attempt.test.ts
@@ -166,132 +166,3 @@ describe("WorkItem.Attempt", () => {
     expect(WorkItem.AttemptId.safeParse(first).success).toBe(true);
   });
 });
-
-describe("WorkItem.CacheKey", () => {
-  const content = () => WorkItem.contentFingerprintOf(contentInputs());
-  const environment = () => WorkItem.environmentFingerprintOf(environmentInputs());
-
-  test("derives from content digest plus a declared deterministic environment subset", () => {
-    const cacheKey = WorkItem.cacheKeyOf(content(), environment(), ["os", "arch", "bunVersion"]);
-    expect(cacheKey.contentDigest).toBe(content().digest);
-    expect(cacheKey.environmentSubset).toEqual(["arch", "bunVersion", "os"]);
-    expect(WorkItem.CacheKey.safeParse(cacheKey).success).toBe(true);
-  });
-
-  test("a declared-absent environment input is not deterministic — selection fails loudly", () => {
-    expect(() => WorkItem.cacheKeyOf(content(), environment(), ["toolVersions"])).toThrow(
-      'cacheKey requires a deterministic environment subset — "toolVersions" is declared absent',
-    );
-  });
-
-  test("the key is derived, never assigned — a tampered key rejects", () => {
-    const cacheKey = WorkItem.cacheKeyOf(content(), environment(), ["os"]);
-    const forged = { ...cacheKey, key: forgeLastChar(cacheKey.key) };
-    expect(refineMessages(WorkItem.CacheKey.safeParse(forged))).toContain(
-      "cacheKey is derived from its components, never assigned",
-    );
-  });
-
-  test("a cacheKey is not a row key — it carries no attempt identity field", () => {
-    expect(Object.keys(WorkItem.cacheKeyOf(content(), environment(), ["os"]))).toEqual([
-      "contentDigest",
-      "environmentSubset",
-      "environmentSubsetDigest",
-      "key",
-    ]);
-  });
-});
-
-describe("WorkItem.ReplayKey", () => {
-  const digest = WorkItem.canonicalDigest("replay");
-
-  function replayKeyOf(overrides: Record<string, unknown> = {}): Record<string, unknown> {
-    return {
-      archivedRange: {
-        streamId: "work:wi_000000000001",
-        fromSeq: 1,
-        toSeq: 5,
-        integrityDigest: digest,
-      },
-      environmentDigest: digest,
-      schemaVersions: { policyKernel: 1 },
-      upcastVersions: { workItem: 1 },
-      nondeterminismManifestDigest: digest,
-      ...overrides,
-    };
-  }
-
-  test("binds an immutable archived range — an inverted range rejects", () => {
-    expect(WorkItem.ReplayKey.safeParse(replayKeyOf()).success).toBe(true);
-    const inverted = WorkItem.ReplayKey.safeParse(
-      replayKeyOf({
-        archivedRange: { streamId: "work:wi", fromSeq: 5, toSeq: 1, integrityDigest: digest },
-      }),
-    );
-    expect(refineMessages(inverted)).toContain("an archived range cannot end before it starts");
-  });
-
-  test("replay-of-record only — the binding has no content-fingerprint slot", () => {
-    const withContent = WorkItem.ReplayKey.safeParse(replayKeyOf({ contentDigest: digest }));
-    expect(withContent.success).toBe(false);
-  });
-});
-
-describe("WorkItem.NondeterminismManifest", () => {
-  const categories = WorkItem.NondeterminismCategory.options;
-
-  function fullCoverage(): Record<string, unknown> {
-    return {
-      recorded: [
-        { category: "clock", identifier: "ref:clock/system", value: 1_700_000_000_000 },
-        { category: "random", identifier: "ref:random/seed", value: "seed-1" },
-      ],
-      absent: categories
-        .filter((category) => category !== "clock" && category !== "random")
-        .map((category) => ({ category, reason: `${category} input was not consumed` })),
-    };
-  }
-
-  test("full coverage — recorded plus absent-with-reason — parses", () => {
-    expect(WorkItem.NondeterminismManifest.safeParse(fullCoverage()).success).toBe(true);
-  });
-
-  test("a category that is neither recorded nor declared absent fails loudly", () => {
-    const manifest = fullCoverage();
-    manifest.absent = (manifest.absent as { category: string }[]).filter(
-      (entry) => entry.category !== "network",
-    );
-    const rejected = WorkItem.NondeterminismManifest.safeParse(manifest);
-    expect(refineMessages(rejected)).toContain(
-      'nondeterminism input "network" is neither recorded nor declared absent with a reason — missing manifest input fails loudly',
-    );
-  });
-
-  test("an absence without a reason is not a declaration", () => {
-    const manifest = fullCoverage();
-    (manifest.absent as Record<string, unknown>[])[0] = { category: "time_zone", reason: "" };
-    expect(WorkItem.NondeterminismManifest.safeParse(manifest).success).toBe(false);
-  });
-
-  test("a category cannot be both recorded and declared absent", () => {
-    const manifest = fullCoverage();
-    (manifest.absent as Record<string, unknown>[]).push({
-      category: "clock",
-      reason: "duplicate declaration",
-    });
-    const rejected = WorkItem.NondeterminismManifest.safeParse(manifest);
-    expect(refineMessages(rejected)).toContain(
-      'nondeterminism input "clock" cannot be both recorded and declared absent',
-    );
-  });
-
-  test("a raw secret is rejected as a recording identifier", () => {
-    const manifest = fullCoverage();
-    (manifest.recorded as Record<string, unknown>[]).push({
-      category: "environment",
-      identifier: "MY_SECRET=hunter2",
-      value: null,
-    });
-    expect(WorkItem.NondeterminismManifest.safeParse(manifest).success).toBe(false);
-  });
-});
