+++ packages/protocol/src/work-item/attempt.ts
@@ -19,10 +19,6 @@ import { z } from "zod";
  *     repeat across attempts.
  *   - `reusedFromAttemptId` records a cache hit: the hit CREATES this new
  *     attempt and points at the attempt whose result it reuses.
- *
- * `CacheKey`, `ReplayKey`, and `NondeterminismManifest` ship as dormant
- * vocabulary in C2 (schema only, no lookup implementation) — see
- * docs/implementation-status.md.
  */
 
 // ---------------------------------------------------------------------------
@@ -118,19 +114,9 @@ const JsonValue: z.ZodType<JsonValue> = z.lazy(() =>
  * LISTED with a reason — never silently empty, never omitted.
  */
 const AbsentInput = z.object({ absent: z.literal(true), reason: z.string().min(1) }).strict();
-type AbsentInput = z.infer<typeof AbsentInput>;
 
 function declared<Value extends z.ZodTypeAny>(value: Value) {
   return z.union([value, AbsentInput]);
-}
-
-function isAbsentInput(value: unknown): value is AbsentInput {
-  return (
-    typeof value === "object" &&
-    value !== null &&
-    (value as { absent?: unknown }).absent === true &&
-    typeof (value as { reason?: unknown }).reason === "string"
-  );
 }
 
 const VersionMap = z.record(z.string().min(1), z.union([z.string().min(1), z.number().int()]));
@@ -322,213 +308,3 @@ export const AttemptTerminal = z
   })
   .strict();
 export type AttemptTerminal = z.infer<typeof AttemptTerminal>;
-
-// ---------------------------------------------------------------------------
-// cacheKey — dormant vocabulary (schema only in C2)
-// ---------------------------------------------------------------------------
-
-const environmentInputFields = Object.keys(EnvironmentFingerprintInputs.shape) as [
-  keyof typeof EnvironmentFingerprintInputs.shape,
-  ...(keyof typeof EnvironmentFingerprintInputs.shape)[],
-];
-
-/** Field names of EnvironmentFingerprintInputs eligible for a cacheKey subset. */
-export const EnvironmentSubsetField = z.enum(environmentInputFields);
-export type EnvironmentSubsetField = z.infer<typeof EnvironmentSubsetField>;
-
-/**
- * Explicit cache lookup key: content fingerprint digest plus a DECLARED
- * deterministic environment subset. Never a row key — it carries no
- * attemptId and `key` is derived (a tampered key fails the refine). A hit
- * creates a NEW attempt and records `reusedFromAttemptId` on it.
- */
-export const CacheKey = z
-  .object({
-    contentDigest: Digest,
-    /** Declared deterministic environment subset (sorted, unique field names). */
-    environmentSubset: z.array(EnvironmentSubsetField).min(1),
-    environmentSubsetDigest: Digest,
-    key: Digest,
-  })
-  .strict()
-  .superRefine((cacheKey, ctx) => {
-    const sorted = [...new Set(cacheKey.environmentSubset)].sort();
-    if (sorted.join(",") !== cacheKey.environmentSubset.join(",")) {
-      ctx.addIssue({
-        code: "custom",
-        message: "environment subset must be sorted and unique",
-        path: ["environmentSubset"],
-      });
-      return;
-    }
-    const derived = canonicalDigest({
-      contentDigest: cacheKey.contentDigest,
-      environmentSubset: cacheKey.environmentSubset,
-      environmentSubsetDigest: cacheKey.environmentSubsetDigest,
-    });
-    if (cacheKey.key !== derived) {
-      ctx.addIssue({
-        code: "custom",
-        message: "cacheKey is derived from its components, never assigned",
-        path: ["key"],
-      });
-    }
-  });
-export type CacheKey = z.infer<typeof CacheKey>;
-
-/**
- * Derives the cache lookup key. Every subset field must be PRESENT in the
- * environment inputs — a declared-absent input is not deterministic and
- * fails loudly.
- */
-export function cacheKeyOf(
-  content: ContentFingerprint,
-  environment: EnvironmentFingerprint,
-  subset: readonly EnvironmentSubsetField[],
-): CacheKey {
-  const fields = [...new Set(subset)].sort();
-  if (fields.length === 0) throw new Error("cacheKey requires a declared environment subset");
-  const picked: Record<string, unknown> = {};
-  for (const field of fields) {
-    const input = environment.inputs[field];
-    if (isAbsentInput(input)) {
-      throw new Error(
-        `cacheKey requires a deterministic environment subset — "${field}" is declared absent (${input.reason})`,
-      );
-    }
-    picked[field] = input;
-  }
-  const environmentSubsetDigest = canonicalDigest(picked);
-  return CacheKey.parse({
-    contentDigest: content.digest,
-    environmentSubset: fields,
-    environmentSubsetDigest,
-    key: canonicalDigest({
-      contentDigest: content.digest,
-      environmentSubset: fields,
-      environmentSubsetDigest,
-    }),
-  });
-}
-
-// ---------------------------------------------------------------------------
-// replayKey — dormant vocabulary (schema only in C2)
-// ---------------------------------------------------------------------------
-
-/**
- * Replay-of-record binding ONLY: an immutable archived range/cassette,
- * the environment fingerprint digest, schema/upcast versions, and the
- * nondeterminism-manifest digest. It carries NO content fingerprint, so it
- * structurally cannot express a cache lookup — never a cache key. An
- * incompatible environment or upcast fails loudly at the (phase D) replay
- * boundary; a what-if/fork is a separately labeled NEW attempt.
- */
-export const ReplayKey = z
-  .object({
-    archivedRange: z
-      .object({
-        streamId: z.string().min(1),
-        fromSeq: z.number().int().positive(),
-        toSeq: z.number().int().positive(),
-        /** Integrity hash of the archived range identity. */
-        integrityDigest: Digest,
-      })
-      .strict(),
-    environmentDigest: Digest,
-    schemaVersions: VersionMap,
-    upcastVersions: VersionMap,
-    nondeterminismManifestDigest: Digest,
-  })
-  .strict()
-  .superRefine((replayKey, ctx) => {
-    if (replayKey.archivedRange.toSeq < replayKey.archivedRange.fromSeq) {
-      ctx.addIssue({
-        code: "custom",
-        message: "an archived range cannot end before it starts",
-        path: ["archivedRange", "toSeq"],
-      });
-    }
-  });
-export type ReplayKey = z.infer<typeof ReplayKey>;
-
-// ---------------------------------------------------------------------------
-// nondeterminism manifest — dormant vocabulary (schema only in C2)
-// ---------------------------------------------------------------------------
-
-/**
- * The issue's enumerated nondeterminism categories: consumed clocks/time
- * zones, random seeds/bytes, model sampling/output/provider request ID,
- * network/tool/device responses, ordering/concurrency choices, generated
- * IDs, environment reads, and human/source inputs.
- */
-export const NondeterminismCategory = z.enum([
-  "clock",
-  "time_zone",
-  "random",
-  "model",
-  "network",
-  "tool",
-  "device",
-  "ordering",
-  "generated_id",
-  "environment",
-  "human",
-  "source",
-]);
-export type NondeterminismCategory = z.infer<typeof NondeterminismCategory>;
-
-const ManifestRecording = z
-  .object({
-    category: NondeterminismCategory,
-    /** Redacted provenance/version reference — secrets never appear raw. */
-    identifier: ReferenceId,
-    value: JsonValue,
-  })
-  .strict();
-
-const ManifestAbsence = z
-  .object({
-    category: NondeterminismCategory,
-    reason: z.string().min(1),
-  })
-  .strict();
-
-/**
- * Fail-loud coverage: EVERY category is either recorded or declared absent
- * with a reason. A category that is neither recorded nor listed rejects —
- * declared-but-absent inputs are listed, never silently omitted.
- */
-export const NondeterminismManifest = z
-  .object({
-    recorded: z.array(ManifestRecording),
-    absent: z.array(ManifestAbsence),
-  })
-  .strict()
-  .superRefine((manifest, ctx) => {
-    const recorded = new Set(manifest.recorded.map((entry) => entry.category));
-    const absent = new Set(manifest.absent.map((entry) => entry.category));
-    if (absent.size !== manifest.absent.length) {
-      ctx.addIssue({
-        code: "custom",
-        message: "a category is declared absent more than once",
-        path: ["absent"],
-      });
-    }
-    for (const category of NondeterminismCategory.options) {
-      if (!recorded.has(category) && !absent.has(category)) {
-        ctx.addIssue({
-          code: "custom",
-          message: `nondeterminism input "${category}" is neither recorded nor declared absent with a reason — missing manifest input fails loudly`,
-          path: ["absent"],
-        });
-      }
-      if (recorded.has(category) && absent.has(category)) {
-        ctx.addIssue({
-          code: "custom",
-          message: `nondeterminism input "${category}" cannot be both recorded and declared absent`,
-          path: ["absent"],
-        });
-      }
-    }
-  });
-export type NondeterminismManifest = z.infer<typeof NondeterminismManifest>;
