--- packages/protocol/test/command/schema.test.ts
@@ -1,179 +1,0 @@
-import { describe, expect, test } from "bun:test";
-import { Command, Ledger, Policy } from "../../src/index.js";
-
-const actor: Command.ActorContext = {
-  kind: "internal_worker",
-  actorId: "worker-1",
-  agentName: "agent",
-  sessionId: "session-1",
-  runId: "run-1",
-};
-
-const target: Command.Target = {
-  kind: "resident",
-  id: "resident-main",
-  sessionId: "resident-session",
-  parentSessionId: "parent-session",
-};
-
-describe("Command protocol schemas", () => {
-  test("Input accepts the public dispatch envelope", () => {
-    const parsed = Command.Input.parse({
-      action: "resident.ask",
-      target,
-      payload: { message: "hello" },
-      wait: true,
-      timeoutMs: 1_000,
-      correlation: "corr-1",
-      idempotencyKey: "idem-1",
-    });
-
-    expect(parsed.action).toBe("resident.ask");
-    expect(parsed.target.kind).toBe("resident");
-    expect(parsed.target.parentSessionId).toBe("parent-session");
-    expect(parsed.wait).toBe(true);
-  });
-
-  test("Target accepts connector endpoint selectors for worker admission", () => {
-    const parsed = Command.Target.parse({
-      kind: "worker",
-      name: "coder",
-      endpointId: "endpoint:install-app-codex",
-      connectorInstallationId: "install:app.example-worker",
-    });
-
-    expect(parsed.endpointId).toBe("endpoint:install-app-codex");
-    expect(parsed.connectorInstallationId).toBe("install:app.example-worker");
-  });
-
-  test("Target rejects executorKind as a public dispatch selector", () => {
-    const parsed = Command.Target.safeParse({
-      kind: "worker",
-      name: "coder",
-      executorKind: "connector_endpoint",
-    });
-
-    expect(parsed.success).toBe(false);
-  });
-
-  test("Input rejects actor and runtime-only fields", () => {
-    for (const field of [
-      "actor",
-      "sessionId",
-      "runId",
-      "workspaceRoot",
-      "model",
-      "provider",
-      "budget",
-      "tools",
-      "permissions",
-      "owner",
-      "ownerId",
-      "dispatchOwners",
-      "routeOwner",
-      "resolveOwner",
-    ]) {
-      expect(
-        Command.Input.safeParse({
-          action: "worker.spawn",
-          target: { kind: "worker" },
-          [field]: field === "actor" ? { actorId: "fake" } : "fake",
-        }).success,
-      ).toBe(false);
-    }
-  });
-
-  test("ActorContext accepts runtime-derived actor shapes", () => {
-    expect(Command.ActorContext.parse(actor).kind).toBe("internal_worker");
-    expect(Command.ActorContext.parse({ kind: "resident", actorId: "resident-main" }).kind).toBe(
-      "resident",
-    );
-    expect(Command.ActorContext.parse({ kind: "system", actorId: "scheduler" }).kind).toBe(
-      "system",
-    );
-    expect(
-      Command.ActorContext.parse({ kind: "unknown", actorId: "unknown", reason: "missing" }).kind,
-    ).toBe("unknown");
-  });
-
-  test("ActorContext refuses the retired command-local kind values (#498 A2)", () => {
-    // "worker"/"user" left the wire vocabulary: new writes are canonical
-    // Actor.Kind only; the historical values survive solely as the persisted
-    // verdict-fact upcast below.
-    for (const kind of ["worker", "user"]) {
-      expect(Command.ActorContext.safeParse({ kind, actorId: "actor-1" }).success).toBe(false);
-    }
-  });
-
-  test("persisted command verdict facts upcast retired actor kinds on read (#498 A2)", () => {
-    const base = {
-      policyId: "dispatch.default",
-      reason: "ok",
-      action: "worker.spawn",
-      targetKind: "worker",
-    };
-
-    // Old persisted bytes decode to the canonical vocabulary…
-    expect(
-      Ledger.CommandAuthorized.parse({ ...base, verdict: "allow", actorKind: "worker" }).actorKind,
-    ).toBe("internal_worker");
-    expect(
-      Ledger.CommandDenied.parse({ ...base, verdict: "deny", actorKind: "user" }).actorKind,
-    ).toBe("human");
-
-    // …and canonical values pass through unchanged.
-    expect(
-      Ledger.CommandAuthorized.parse({ ...base, verdict: "allow", actorKind: "resident" })
-        .actorKind,
-    ).toBe("resident");
-    expect(
-      Ledger.CommandAuthorized.parse({
-        ...base,
-        verdict: "allow",
-        actorKind: "internal_worker",
-      }).actorKind,
-    ).toBe("internal_worker");
-  });
-
-  test("Request carries canonical runtime metadata", () => {
-    const command = Command.Request.parse({
-      action: "resident.ask",
-      target,
-      payload: { message: "hello" },
-      dispatchId: "dispatch-1",
-      actor,
-      traceId: "trace-1",
-      sessionId: "session-1",
-      runId: "run-1",
-      workspaceRoot: "/workspace",
-      submittedAt: Date.now(),
-    });
-
-    expect(command.actor.actorId).toBe("worker-1");
-  });
-
-  test("Policy.Resource accepts dispatch descriptors for policy audit", () => {
-    const descriptor = Policy.Resource.Descriptor.parse({
-      id: "dispatch:resident.ask",
-      kind: "dispatch",
-      labels: ["dispatch.resident"],
-      capabilities: ["resident.ask"],
-      effects: ["message.deliver"],
-      source: { type: "runtime", runtimeId: "dispatch-runtime" },
-    });
-
-    expect(descriptor.kind).toBe("dispatch");
-  });
-
-  test("Policy timing and point expose the canonical dispatch authority gate", () => {
-    expect(Policy.Timing.DISPATCH_AUTHORIZE).toBe("dispatch.authorize");
-
-    const point = Policy.PolicyPoint.Registry["dispatch.action.pre"];
-    expect(point.phase).toBe("pre");
-    expect(point.defaultFailPolicy).toBe("fail-closed");
-    expect(point.sideEffectBoundary).toBe(true);
-    expect(point.requiredContext).toEqual(["actor", "dispatchId", "action", "target"]);
-    expect(point.allowedEffects).toEqual(["audit.annotate", "run.abort"]);
-    expect(Policy.PolicyPoint.Id.parse("dispatch.action.pre")).toBe("dispatch.action.pre");
-  });
-});
